Update dbase-rs to 0.1.1
The various `iter_shapes_as` `iter_shapes_and_records_as`
now take `&mut self` instead of `self` to be able
to use `dbase::RecordIterator`.

Move the purely shapefile related reader in a new
`ShapeReader` struct and use it in the `Reader` struct.
`ShapeReader` may be useful for projects wich don't care about dbf

Same thing for writing
`ShapeWriter is a new struct that handle the writing of the shp and shx.

The `Reader` has a handle to a `ShapeWriter` and `dbase::TableWriter`.