Write shape one by one
Add `write_shape` to the ShapeWriter:
this method allows the user to write one shape at a time.
This change forces the type of the destination to implement
`std::io::Seek`

Change the `write_shapes` to take any type that implements
`IntoIterator<Item=&ShapeType>` which makes the function more generic.

Contains a fix for the Multipatch shapetype which wasn't
registered as a shapetype with Z coordinates